// client/src/interfaces/total-stats.ts
export interface TotalStats {
	currentMonthOnlineSellsTotalRevenueAmount: number;
	currentMonthOfflineSellsTotalRevenueAmount: number;
	currentMonthTotalRevenueAmount: number;
	lastMonthOnlineSellsTotalRevenueAmount: number;
	lastMonthOfflineSellsTotalRevenueAmount: number;
	lastMonthTotalRevenueAmount: number;
	currentMonthOnlineSellsTotalProductsAmount: number;
	currentMonthOfflineSellsTotalProductsAmount: number;
	currentMonthTotalProductsAmount: number;
	lastMonthOnlineSellsTotalProductsAmount: number;
	lastMonthOfflineSellsTotalProductsAmount: number;
	lastMonthTotalProductsAmount: number;
	currentMonthOnlineSellersTotalWage: number;
	currentMonthOfflineSellersTotalWage: number;
	currentMonthTotalWage: number;
	lastMonthOnlineSellersTotalWage: number;
	lastMonthOfflineSellersTotalWage: number;
	lastMonthTotalWage: number;
}